move and fix tests
